// src/app/api/model/getStandingInstructionsStandingInstructionIdResponse.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements. See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership. The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GetToClientStandingInstructionSwagger } from './getToClientStandingInstructionSwagger';
import { GetStatusStandingInstructionSwagger } from './getStatusStandingInstructionSwagger';
import { GetToAccountStandingInstructionSwagger } from './getToAccountStandingInstructionSwagger';
import { GetFromAccountTypeStandingInstructionSwagger } from './getFromAccountTypeStandingInstructionSwagger';
import { GetRecurrenceFrequencyStandingInstructionSwagger } from './getRecurrenceFrequencyStandingInstructionSwagger';
import { GetRecurrenceTypeStandingInstructionSwagger } from './getRecurrenceTypeStandingInstructionSwagger';
import { GetFromOfficeStandingInstructionSwagger } from './getFromOfficeStandingInstructionSwagger';
import { GetToAccountTypeStandingInstructionSwagger } from './getToAccountTypeStandingInstructionSwagger';
import { GetFromAccountStandingInstructionSwagger } from './getFromAccountStandingInstructionSwagger';
import { GetPriorityStandingInstructionSwagger } from './getPriorityStandingInstructionSwagger';
import { GetToOfficeStandingInstructionSwagger } from './getToOfficeStandingInstructionSwagger';
import { GetFromClientStandingInstructionSwagger } from './getFromClientStandingInstructionSwagger';
import { GetTransferTypeStandingInstructionSwagger } from './getTransferTypeStandingInstructionSwagger';
import { GetInstructionTypeStandingInstructionSwagger } from './getInstructionTypeStandingInstructionSwagger';


/**
 * GetStandingInstructionsStandingInstructionIdResponse
 */
export interface GetStandingInstructionsStandingInstructionIdResponse { 
    accountDetailId?: number;
    amount?: number;
    fromAccount?: GetFromAccountStandingInstructionSwagger;
    fromAccountType?: GetFromAccountTypeStandingInstructionSwagger;
    fromClient?: GetFromClientStandingInstructionSwagger;
    fromOffice?: GetFromOfficeStandingInstructionSwagger;
    id?: number;
    instructionType?: GetInstructionTypeStandingInstructionSwagger;
    name?: string;
    priority?: GetPriorityStandingInstructionSwagger;
    recurrenceFrequency?: GetRecurrenceFrequencyStandingInstructionSwagger;
    recurrenceInterval?: number;
    recurrenceOnMonthDay?: string;
    recurrenceType?: GetRecurrenceTypeStandingInstructionSwagger;
    status?: GetStatusStandingInstructionSwagger;
    toAccount?: GetToAccountStandingInstructionSwagger;
    toAccountType?: GetToAccountTypeStandingInstructionSwagger;
    toClient?: GetToClientStandingInstructionSwagger;
    toOffice?: GetToOfficeStandingInstructionSwagger;
    transferType?: GetTransferTypeStandingInstructionSwagger;
    validFrom?: string;
}